requirements.txt
django-countries
django-countries
boto3==1.9.28
botocore==1.12.28
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0
django-adminplus==0.5
django-admin-bootstrapped==2.5.7
django-allauth==0.36.0
django-cors-headers==2.2.0
django-extensions==2.1.0
django-storages==1.7.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
greenlet==0.4.14
gunicorn==19.8.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mock==2.0.0
msgpack==0.5.6
neovim==0.2.6
oauthlib==2.1.0
pbr==4.2.0
psycopg2==2.7.4
psycopg2-binary==2.7.5
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.8.2
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.11.0
stevedore==1.29.0
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
whitenoise==4.0
django-countries
